requirements.txt
selenium~=4.15.2
pytest~=7.4.3
h11~=0.14.0
pip~=23.3.1
attrs~=23.1.0
wheel~=0.38.4
Jinja2~=3.1.2
setuptools~=65.5.1
decorator~=5.1.1
trio~=0.23.1
outcome~=1.3.0.post0
sniffio~=1.3.0
sortedcontainers~=2.4.0
cffi~=1.16.0
idna~=3.4
execnet~=2.0.2
MarkupSafe~=2.1.3
pluggy~=1.3.0
iniconfig~=2.0.0
certifi~=2023.7.22
urllib3~=2.0.7
pycparser~=2.21
packaging~=23.2
wsproto~=1.2.0
allure-pytest~=2.13.2
